created functional for showing event participants in info

+++ src/main/java/com/webapp/sportmeetingpoint/persistance/NewsRepository.java
@@ -3,6 +3,6 @@ package com.webapp.sportmeetingpoint.persistance;
 import com.webapp.sportmeetingpoint.domain.entities.News;
 import org.springframework.data.repository.CrudRepository;
 
-public interface NewsRepository extends CrudRepository<News, Long> {
+public interface NewsRepository extends CrudRepository<News, Integer> {
     News save(News news);
 }
