+++ src/main/java/com/webapp/sportmeetingpoint/persistance/UserAuthorActivityRepository.java
@@ -3,7 +3,7 @@ package com.webapp.sportmeetingpoint.persistance;
 import com.webapp.sportmeetingpoint.domain.entities.UserAuthorActivity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserAuthorActivityRepository extends CrudRepository<UserAuthorActivity, Long> {
+public interface UserAuthorActivityRepository extends CrudRepository<UserAuthorActivity, Integer> {
 
 //  UserAuthorActivity save(UserAuthorActivity userAuthorActivity);
