+++ src/main/java/com/webapp/sportmeetingpoint/persistance/UserPersonalDataRepository.java
@@ -3,7 +3,7 @@ package com.webapp.sportmeetingpoint.persistance;
 import com.webapp.sportmeetingpoint.domain.entities.UserPersonalData;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserPersonalDataRepository extends CrudRepository<UserPersonalData, Long> {
+public interface UserPersonalDataRepository extends CrudRepository<UserPersonalData, Integer> {
 
   UserPersonalData save(UserPersonalData personalData);
